requirements.txt
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
networkx==2.5.1
numpy==1.20.1
pandas==1.2.4
Pillow==8.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
threadpoolctl==2.1.0
tifffile==2021.4.8
torch==1.8.1+cu102
torchaudio==0.8.1
torchvision==0.9.1+cu102
typing-extensions==3.10.0.0
Werkzeug==1.0.1
